requirements.txt
# Dependencias del Sistema de Adaptación de Contenido

# Core dependencies
openai>=1.3.0
python-dotenv>=0.19.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP requests
requests>=2.31.0
httpx>=0.25.0

# Data handling
python-dateutil>=2.8.0
pydantic>=2.4.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Task Queue
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Optional: Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
python-multipart